Loops support for RDNA1-2 + RDNA3 support

+++ tests/real_kernels/bellmanFordInit_ver1/bellmanFordInit_ver1_dcmpl.cl
@@ -11,7 +11,7 @@ void bellmanFordIter_ver1(uint nEdges, __global uint2 *edges, __global double *w
         if ((double)var1 < (double)1e12) {
             var2 = d[var0.s1];
             var3 = weights[get_global_id(0)];
-            if ((uint)var2 > (uint)((double)var1 + (double)var3)) {
+            if ((double)var2 > (double)((double)var1 + (double)var3)) {
                 d[var0.s1] = (double)var1 + (double)var3;
                 *changed = 1;
             }
